refactor: rename feed_types -> feed
