+++ src/core/data_types/video/factory.rs
@@ -1,4 +1,4 @@
-use crate::core::data_types::{feed_types::*, video::video::Video};
+use crate::core::data_types::{feed::*, video::video::Video};
 use chrono::DateTime;
 
 #[derive(Default)]
